@@ -1,59 +1,48 @@
 {
   "name": "quill-mention",
-  "version": "2.1.4",
+  "version": "2.2.5",
   "description": "@mentions for the Quill rich text editor",
-  "main": "src/quill.mention.js",
-  "repository": "https://github.com/afconsult/quill-mention.git",
+  "main": "dist/quill.mention.csj.js",
+  "module": "dist/quill.mention.esm.js",
+  "repository": "https://github.com/afry/quill-mention.git",
   "author": "Fredrik Sundqvist <fsundqvist@gmail.com>",
   "license": "MIT",
   "files": [
-    "src",
     "dist",
     "LICENSE",
     "README.md"
   ],
-  "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.3",
-    "babel-plugin-transform-object-assign": "^6.22.0",
-    "babel-preset-env": "^1.6.1",
-    "css-loader": "^0.28.8",
-    "eslint": "^4.15.0",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-plugin-import": "^2.8.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-react": "^7.5.1",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "jshint": "^2.9.5",
-    "style-loader": "^0.19.1",
-    "uglifyjs-webpack-plugin": "^1.1.6",
-    "webpack": "4.19.1",
-    "webpack-cli": "^2.0.10",
-    "webpack-dev-server": ">=3.1.11",
-    "webpack-merge": "^4.2.1"
-  },
   "dependencies": {
     "quill": "^1.3.4"
   },
-  "eslintConfig": {
-    "extends": "airbnb",
-    "env": {
-      "browser": true
-    },
-    "rules": {
-      "import/no-extraneous-dependencies": [
-        "error",
-        {
-          "devDependencies": true
-        }
-      ]
-    }
+  "devDependencies": {
+    "@ava/babel": "^0.4.0",
+    "@babel/core": "^7.7.4",
+    "@babel/plugin-transform-object-assign": "^7.7.4",
+    "@babel/preset-env": "^7.7.4",
+    "@babel/register": "^7.7.7",
+    "ava": "^2.4.0",
+    "concurrently": "^5.0.2",
+    "eslint": "^6.8.0",
+    "eslint-config-prettier": "^6.9.0",
+    "eslint-plugin-prettier": "^3.1.2",
+    "prettier": "1.19.1",
+    "rollup": "^1.27.8",
+    "rollup-plugin-babel": "^4.3.3",
+    "rollup-plugin-local-resolve": "^1.0.7",
+    "rollup-plugin-postcss": "^2.0.3",
+    "rollup-plugin-terser": "^5.1.2",
+    "serve": "^11.2.0"
   },
   "scripts": {
-    "start": "webpack-dev-server --hot --open --config webpack.dev.js",
-    "build": "webpack --config webpack.prod.js",
-    "publish": "npm publish"
+    "start": "concurrently \"rollup -c -w\" \"serve docs\"",
+    "build": "rollup -c",
+    "test": "ava"
+  },
+  "ava": {
+    "require": [
+      "@babel/register"
+    ]
   },
   "keywords": [
     "quill",
