@@ -3,7 +3,7 @@ const Keys = {
   ENTER: 13,
   ESCAPE: 27,
   UP: 38,
-  DOWN: 40,
+  DOWN: 40
 };
 
 export default Keys;
